backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.8.5
httpx==0.25.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.3.0